src/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.12
pandas==2.2.3
openpyxl==3.1.5
anyio==4.6.2
httpx==0.27.2
